Refactored for results in a page to use cards generated in the favourite table generation code
