+++ src/aims_ui/static/js/f_helpers/address_attribute_map.mjs
@@ -1,3 +1,5 @@
+// Map a flattened address object to an array of objects representing
+// the attributes of an address
 
 export function getPopulatedAttributeMap(addressObject) {
   // Return an array of objects mapping with cellId: and value:, labelText:
